fixed stack overflow in release build with long filter string
Resolved an issue where using long filter strings in release mode caused a stack overflow.

I'm not sure if this is set correctly, as I don't know much about compiler flags.

This may need to be revised in the future due to the noticeable size(in my case, 80kb) difference compared to the original WinDivert DLL.